Method for fast project navigation
Added a ZSH-function which enables me to open my projects in a fast way.
Because I use PHPStorm, I know there will be a .idea-folder in the project, so
by search for this folder I know the location

If the location is found (@todo errorhandling) I open a finder window, change
the director for the terminal into that directory and fire up PHP Storm with
the project loaded.